Reorganise effects registry by package contributions
