+++ crates/oak_semantic/src/effects.rs
@@ -4,10 +4,8 @@ use aether_syntax::AnyRValue;
 use aether_syntax::RArgument;
 use aether_syntax::RBinaryExpression;
 use aether_syntax::RCall;
-use biome_rowan::AstNode;
 use biome_rowan::AstPtr;
 use biome_rowan::AstSeparatedList;
-use biome_rowan::WalkEvent;
 // Re-exported so consumers building an `AssignBinding` (custom `AssignHandler`s)
 // can name the `name_expr` field's type without depending on oak_core directly.
 pub use oak_core::range::RangedAstPtr;
@@ -16,6 +14,10 @@ use oak_core::syntax_ext::RStringValueExt;
 
 use crate::semantic_index::NseScope;
 use crate::semantic_index::NseTiming;
+
+/// Per-package tables of which functions carry effects. Private data behind the
+/// `lookup`/`annotates` query API below.
+mod contrib;
 
 /// Effects of a call, resolved against the call site.
 #[derive(Debug, Clone, Default)]
@@ -57,6 +59,25 @@ pub struct EffectsHandlers {
     pub attach: Option<&'static dyn EffectHandler<Output = String>>,
     pub source: Option<&'static dyn EffectHandler<Output = Vec<String>>>,
     pub assign: Option<&'static dyn AssignHandler>,
+}
+
+/// Look up the effect handlers of a `(package, function)` pair.
+pub fn lookup(package: &str, function: &str) -> Option<&'static EffectsHandlers> {
+    contrib::REGISTRY
+        .iter()
+        .flat_map(|entries| entries.iter())
+        .find(|entry| entry.package == package && entry.function == function)
+        .map(|entry| &entry.effects)
+}
+
+/// Whether any registry entry annotates `name`. This is the bare-callee front
+/// gate: an unannotated name can't resolve to an effect no matter which provider
+/// wins, so recognition skips resolution entirely.
+pub fn annotates(name: &str) -> bool {
+    contrib::REGISTRY
+        .iter()
+        .flat_map(|entries| entries.iter())
+        .any(|entry| entry.function == name)
 }
 
 /// Resolver for an effect of a call.
@@ -281,107 +302,6 @@ impl EffectHandler for ArgumentsAnnotation {
                 .collect(),
         )
     }
-}
-
-/// Handler for `bquote()`. It quotes its `expr` argument like `quote()`, but a
-/// `.(X)` inside escapes back to evaluation, so `X` is a live sub-expression.
-/// Recognizing `.()` is specific to bquote, so it lives in this handler rather
-/// than in the shared [`ArgumentEffect`] vocabulary.
-#[derive(Debug, Clone, Copy)]
-pub struct BquoteHandler;
-
-impl EffectHandler for BquoteHandler {
-    type Output = ResolvedArgumentEffects;
-
-    fn resolve(&self, call: &RCall, ctx: &CallContext) -> Option<ResolvedArgumentEffects> {
-        // `bquote(expr, where, splice)`: only `expr` (the first positional) is
-        // quoted. The other arguments are ordinary values.
-        let formals = [
-            Formal {
-                name: "expr",
-                position: 0,
-            },
-            Formal {
-                name: "splice",
-                position: 2,
-            },
-        ];
-        let matched = ctx.match_arguments(call, &formals);
-
-        let args = call.arguments().ok()?;
-        let values: Vec<Option<AnyRExpression>> = args
-            .items()
-            .iter()
-            .map(|item| item.ok().and_then(|arg| arg.value()))
-            .collect();
-
-        // `..()` only splices under `splice = TRUE`.
-        // TODO: resolve a dynamic `splice` argument.
-        let splice = matched
-            .iter()
-            .position(|formal| *formal == Some(1))
-            .and_then(|i| values.get(i))
-            .and_then(|value| value.as_ref())
-            .and_then(|value| ctx.resolve_static_bool(value))
-            .unwrap_or(false);
-
-        Some(
-            matched
-                .into_iter()
-                .enumerate()
-                .map(|(i, formal)| {
-                    // Only `expr` (formal 0) is quoted
-                    if formal != Some(0) {
-                        return None;
-                    }
-                    let holes = values
-                        .get(i)
-                        .and_then(|value| value.as_ref())
-                        .map(|expr| unquote_holes(expr, splice))
-                        .unwrap_or_default();
-                    Some(ResolvedArgumentEffect::Quote { holes })
-                })
-                .collect(),
-        )
-    }
-}
-
-/// The unquote holes inside a bquote-quoted expression: the escaped argument of
-/// each `.(foo)` call, plus each `..(foo)` when `splice` is on.
-fn unquote_holes(expr: &AnyRExpression, splice: bool) -> Vec<AnyRExpression> {
-    let mut holes = Vec::new();
-    let mut preorder = expr.syntax().preorder();
-    while let Some(event) = preorder.next() {
-        let WalkEvent::Enter(node) = event else {
-            continue;
-        };
-        let Some(call) = RCall::cast(node) else {
-            continue;
-        };
-        if let Some(hole) = unquote_hole(&call, splice) {
-            holes.push(hole);
-            preorder.skip_subtree();
-        }
-    }
-    holes
-}
-
-/// The escaped expression of a `.(foo)` unquote call, or a `..(foo)` splice
-/// unquote when `splice` is on, or `None` when `call` isn't one. bquote's
-/// unquote operator is the function `.`, and its splice unquote is `..`.
-fn unquote_hole(call: &RCall, splice: bool) -> Option<AnyRExpression> {
-    let AnyRExpression::RIdentifier(func) = call.function().ok()? else {
-        return None;
-    };
-    let is_unquote = match func.name_text().as_str() {
-        "." => true,
-        ".." => splice,
-        _ => false,
-    };
-    if !is_unquote {
-        return None;
-    }
-    call.arguments().ok()?.items().iter().next()?.ok()?.value()
 }
 
 /// Declares how an attach function (`library()`, `require()`) names its package,
