+++ crates/oak_semantic/src/effects/contrib/magrittr.rs
@@ -1,0 +1,5 @@
+use crate::effects::contrib::assign_op;
+use crate::effects::contrib::Entry;
+use crate::effects::TargetAccess::ReadWrite;
+
+pub(crate) static ENTRIES: &[Entry] = &[assign_op!("magrittr", "%<>%", ReadWrite)];
