+++ crates/oak_semantic/src/effects/contrib/rlang.rs
@@ -1,0 +1,11 @@
+use crate::effects::contrib::assign_op;
+use crate::effects::contrib::nse;
+use crate::effects::contrib::Entry;
+use crate::effects::TargetAccess::Write;
+use crate::semantic_index::NseScope::Current;
+use crate::semantic_index::NseTiming::Lazy;
+
+pub(crate) static ENTRIES: &[Entry] = &[
+    assign_op!("rlang", "%<~%", Write),
+    nse!("rlang", "on_load", ("expr", 0, Current, Lazy)),
+];
