+++ crates/oak_semantic/src/effects/contrib/s7.rs
@@ -1,0 +1,5 @@
+use crate::effects::contrib::assign_op;
+use crate::effects::contrib::Entry;
+use crate::effects::TargetAccess::Write;
+
+pub(crate) static ENTRIES: &[Entry] = &[assign_op!("S7", ":=", Write)];
